+++ index.php
@@ -1,7 +1,13 @@
 <?php
 
+require_once("init.php");
+require_once("functions.php");
 require_once("helpers.php");
 require_once("data.php");
+require_once("logic.php");
+
+
+$title = "Дела в порядке";
 
 $page_content = include_template("main.php", [
     "arr_tasks" => $arr_tasks,
